Enum c_certitude::ValidationResult

pub enum ValidationResult {
    Trusted,
    NotTrusted,
    MalformedCertificateInChain,
    UnableToBuildTrustStore,
    ErrorDuringValidation,
    MissingFunctionality,
    UserAuthenticationRequired,
    MalformedHostname,
}

Possible results from attempting to validate a certificate chain.

When attempting to validate a certificate chain, in addition to the two "successful" results (ValidationResult::Trusted and ValidationResult::NotTrusted), there are numerous possible error conditions. This enum allows for expressing those error conditions.

Note that due to the vagaries of the system libraries, it is possible that a misleading error may be generated: for example, the hostname may be malformed but in a manner that does not immediate generate a ValidationResult::MalformedHostname result. That's unfortunate, but there is relatively little that can be done about that in the absence of clearer system APIs.

Variants

Trusted

NotTrusted

MalformedCertificateInChain

UnableToBuildTrustStore

ErrorDuringValidation

MissingFunctionality

UserAuthenticationRequired

MalformedHostname