Function c_certitude::validate_cert_chain

pub extern fn validate_cert_chain(encoded_certs: *mut *const u8, cert_sizes: *mut usize, cert_count: usize, hostname: *const c_char) -> u32

A C-ABI compatible version of the cert validation function. It does some work to transform the data, and then just calls the cert validation function from certitude.

Here are the rules for calling this from C.

encoded_certs is an array of pointers to DER-encoded representations of the certificates in the certificate chain. Neither the pointer array nor the DER-encoded representations of the certs are null-terminated.

cert_sizes is an array of lengths for each DER-encoded cert. This array has the exact same length as the array in encoded_certs. These lengths must not include any null-terminators that may be at the end of the DER-encoded certs.

cert_count is the number of elements in both the encoded_certs and cert_sizes array.

hostname is the UTF-8 encoded hostname to validate the cert chain against. This string must be null-terminated.

Rust takes ownership of none of these objects: where necessary it makes its own copies, but generally speaking it expects all of the data passed to be it to remain valid for the duration of the function call. The caller retains ownership, however, and is responsible for freeing all the relevant data.

The C header for this function is:

extern uint32_t validate_cert_chain(uint8_t **encoded_certs,
                                    size_t *cert_sizes,
                                    size_t cert_count,
                                    const char *hostname);